// api/api/models/CreateProductDto.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CreateProductDto = {
    name: string;
    description?: string;
    image: string;
    quantity: number;
    category: string;
    fat: number;
    carbohydrates: number;
    protein: number;
    weight: number;
    price: string;
    expirationDate: string;
}